requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
frozenlist==1.3.3
idna==3.4
langchain==0.0.190
markdownify==0.11.6
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
tenacity==8.2.2
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
